+++ src/app/dataset-view/additional-components/metadata-component/components/set-transform/components/engine-section/engine-section.component.spec.ts
@@ -12,6 +12,7 @@ import { SharedTestModule } from "src/app/common/modules/shared-test.module";
 import { EngineSelectComponent } from "./components/engine-select/engine-select.component";
 import { EngineDesc } from "src/app/api/kamu.graphql.interface";
 import { findElementByDataTestId } from "src/app/common/helpers/base-test.helpers.spec";
+import AppValues from "src/app/common/values/app.values";
 
 describe("EngineSectionComponent", () => {
     let component: EngineSectionComponent;
@@ -39,8 +40,8 @@ describe("EngineSectionComponent", () => {
 
     [
         { testCase: "spark", expected: mockEngines.data.knownEngines[0] },
-        { testCase: "datafusion", expected: mockEngines.data.knownEngines[1] },
-        { testCase: "flink", expected: mockEngines.data.knownEngines[2] },
+        { testCase: "flink", expected: mockEngines.data.knownEngines[1] },
+        { testCase: "datafusion", expected: mockEngines.data.knownEngines[2] },
         { testCase: "risingwave", expected: mockEngines.data.knownEngines[3] },
     ].forEach((item: { testCase: string; expected: EngineDesc }) => {
         it(`should check init ${item.testCase} engine and image`, fakeAsync(() => {
@@ -57,11 +58,23 @@ describe("EngineSectionComponent", () => {
 
     it("should check init engine and image when currentSetTransformEvent is not null", fakeAsync(() => {
         component.currentSetTransformEvent = mockSetPollingSourceEvent.preprocess;
+        component.selectedEngine = "spark";
         fixture.detectChanges();
         component.ngOnInit();
         tick();
         if (mockSetPollingSourceEvent.preprocess)
             expect(component.selectedEngine).toBe(mockSetPollingSourceEvent.preprocess.engine);
+        flush();
+    }));
+
+    it("should check init engine and image when currentSetTransformEvent is null", fakeAsync(() => {
+        component.currentSetTransformEvent = null;
+        component.selectedEngine = "";
+        fixture.detectChanges();
+        component.ngOnInit();
+        tick();
+
+        expect(component.selectedEngine).toBe(AppValues.DEFAULT_ENGINE_NAME);
         flush();
     }));
 });